llm/journey_icons/icon_matcher/pyproject.toml
[tool.poetry]
name = "icon-matcher"
version = "0.1.0"
description = ""
authors = ["mathatan <mathatan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.2"
langchain-openai = "^0.2.2"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
